tests/aws/non_context_aware/log_validation_rules/test_ensure_cloudwatch_log_groups_specify_retention_days_rule.py

# import unittest
#
# from cloudrail.knowledge.context.environment_context import EnvironmentContext
# from cloudrail.knowledge.rules.base_rule import RuleResultType
# from tests.rule_test_utils import create_empty_entity
#
#
# class TestEnsureCloudWatchLogGroupsRetentionUsageRule(unittest.TestCase):
#     def setUp(self):
#         self.rule = EnsureCloudWatchLogGroupsRetentionUsageRule()
#
#     def test_non_car_cw_log_group_no_retention_fail(self):
#         # Arrange
#         context = EnvironmentContext()
#         # Act
#         result = self.rule.run(context, {})
#         # Assert
#         self.assertEqual(RuleResultType.FAILED, result.status)
#         self.assertEqual(1, len(result.issues))
#
#     def test_non_car_cw_log_group_no_retention_pass(self):
#         # Arrange
#         context = EnvironmentContext()
#         # Act
#         result = self.rule.run(context, {})
#         # Assert
#         self.assertEqual(RuleResultType.SUCCESS, result.status)
#         self.assertEqual(0, len(result.issues))
